vagrant snapshot back no longer fails if the VM is in "aborted" state

+++ lib/vagrant-vbox-snapshot/commands/back.rb
@@ -22,7 +22,7 @@ module VagrantPlugins
           with_target_vms(argv, single_target: true) do |machine|
             check_runnable_on(machine)
 
-            if machine.state.id != :poweroff and machine.state.id != :saved
+            if machine.state.id != :poweroff and machine.state.id != :saved and machine.state.id != :aborted
               machine.provider.driver.execute("controlvm", machine.id, "poweroff")
             end
